resolved many-to-many cascading issue, added javadocs

+++ src/main/java/com/cmpe275/controller/PlayerController.java
@@ -3,43 +3,68 @@ package com.cmpe275.controller;
 import com.cmpe275.domain.Address;
 import com.cmpe275.domain.Player;
 import com.cmpe275.domain.Sponsor;
-import com.cmpe275.exceptions.BadRequestException;
-import com.cmpe275.exceptions.ResourceNotFoundException;
 import com.cmpe275.service.PlayerService;
 import com.cmpe275.service.SponsorService;
 import io.swagger.annotations.Api;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 /**
  *
  * Created by arunabh.shrivastava on 11/8/2017.
+ *
+ * Player Controller to perform CRUD operations on Player
  */
 
 
 @RestController
 @Api(value = "Player management endpoint", description = "This API performs CRUD operations on Players.")
-@RequestMapping(value = "/api/v1/player")
+@RequestMapping(value = "/player")
 public class PlayerController {
 
     private final PlayerService playerService;
     private final SponsorService sponsorService;
 
+    /**
+     * Instantiates a new PlayerController
+     *
+     * @param playerService         Player Service to perform actions on player object
+     * @param sponsorService        Sponsor Service to perform actions on sponsor object
+     */
     @Autowired
     public PlayerController(PlayerService playerService, SponsorService sponsorService) {
         this.playerService = playerService;
         this.sponsorService = sponsorService;
     }
 
+
+    /**
+     * Creates a new Player
+     *
+     * @param firstname         First name of the player
+     * @param lastname          Last name of the player
+     * @param email             Email address of the player
+     * @param description       Description about the player
+     * @param city              City of the player address
+     * @param street            Street name and number of the player address
+     * @param state             State of the player address
+     * @param zip               Zip Code of the player address
+     * @param sponsorId         Sponsor id of the sponsor
+     *
+     * @return A {@link Player} object on success, or a HttpStatus.BAD_REQUEST on invalid request
+     *
+     */
     @PostMapping
-    public Player create(@RequestParam(value = "firstname") String firstname, @RequestParam(value = "lastname")
+    public ResponseEntity<?> create(@RequestParam(value = "firstname") String firstname, @RequestParam(value = "lastname")
             String lastname, @RequestParam(value = "email") String email,
-                         @RequestParam(value = "description", required = false) String description,
-                         @RequestParam(value = "city", required = false) String city,
-                         @RequestParam(value = "street", required = false) String street,
-                         @RequestParam(value = "state", required = false) String state,
-                         @RequestParam(value = "zip", required = false) Integer zip,
-                         @RequestParam(value = "sponsor", required = false) Long sponsorId) {
+                                 @RequestParam(value = "description", required = false) String description,
+                                 @RequestParam(value = "city", required = false) String city,
+                                 @RequestParam(value = "street", required = false) String street,
+                                 @RequestParam(value = "state", required = false) String state,
+                                 @RequestParam(value = "zip", required = false) Integer zip,
+                                 @RequestParam(value = "sponsor", required = false) Long sponsorId) {
 
         Address address = null ;
         Sponsor sponsor = null;
@@ -52,7 +77,7 @@ public class PlayerController {
             sponsor = sponsorService.get(sponsorId);
 
             if(sponsor == null){
-                throw new BadRequestException("Invalid Sponsor ID");
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
             }
         }
 
@@ -60,24 +85,47 @@ public class PlayerController {
         player = playerService.create(player);
 
         if(player == null){
-            throw new BadRequestException("Bad Request");
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
-
-        return player;
+        return new ResponseEntity<Object>(player, HttpStatus.OK);
     }
 
+    /**
+     * Gets a Player
+     *
+     * @param id         Id of the player
+     *
+     * @return A {@link Player} object on success, or a HttpStatus.NOT_FOUND if player doesn't exists
+     */
     @GetMapping(value = "/{id}")
-    public Player get(@PathVariable(value = "id") Long id){
+    public ResponseEntity<?> get(@PathVariable(value = "id") Long id){
 
         Player player = playerService.get(id);
         if(player == null){
-            throw new ResourceNotFoundException("Player not found");
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        return player;
+        return new ResponseEntity<Object>(player, HttpStatus.OK);
     }
 
+    /**
+     * Updates a Player
+     *
+     * @param id                Id of the player
+     * @param firstname         First name of the player
+     * @param lastname          Last name of the player
+     * @param email             Email address of the player
+     * @param description       Description about the player
+     * @param city              City of the player address
+     * @param street            Street name and number of the player address
+     * @param state             State of the player address
+     * @param zip               Zip Code of the player address
+     * @param sponsorId         Sponsor id of the sponsor
+     *
+     * @return A {@link Player} object on success, or HttpStatus.NOT_FOUND if player doesn't exists or
+     * HttpStatus.BAD_REQUEST on invalid request
+     */
     @PutMapping(value = "/{id}")
-    public Player update(@PathVariable("id") Long id, @RequestParam(value = "firstname") String firstname,
+    public ResponseEntity<?> update(@PathVariable("id") Long id, @RequestParam(value = "firstname") String firstname,
                          @RequestParam(value = "lastname") String lastname, @RequestParam(value = "email") String email,
                          @RequestParam(value = "description", required = false) String description,
                          @RequestParam(value = "city", required = false) String city,
@@ -88,7 +136,7 @@ public class PlayerController {
 
         Player player = playerService.get(id);
         if(player == null){
-            throw new ResourceNotFoundException("Player not found");
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
 
         Address address = null;
@@ -101,20 +149,39 @@ public class PlayerController {
         if(sponsorId != null){
             sponsor = sponsorService.get(sponsorId);
             if(sponsor == null){
-                throw new BadRequestException("Invalid Sponsor ID");
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
             }
         }
 
         Player newPlayerInfo = new Player(id, firstname, lastname, email,description, address, sponsor);
         player = playerService.update(newPlayerInfo);
         if(player == null){
-            throw new BadRequestException("Bad Request");
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
-        return player;
+        return new ResponseEntity<Object>(player, HttpStatus.OK);
     }
 
+    /**
+     * Deletes a player
+     *
+     * @param id         Id of the player
+     *
+     * @return A {@link Player} object on success, or HttpStatus.NOT_FOUND if player doesn't exists or
+     * HttpStatus.BAD_REQUEST on invalid request
+     */
     @DeleteMapping(value = "/{id}")
-    public Player delete(@PathVariable("id") Long id){
-        return playerService.delete(id);
+    public ResponseEntity<?> delete(@PathVariable("id") Long id){
+
+        Player player = playerService.get(id);
+        if(player == null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        try {
+            playerService.delete(id);
+            return new ResponseEntity<Object>(player, HttpStatus.OK);
+        }catch (Exception e){
+            e.printStackTrace();
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
     }
 }
